LPS-76524 move workspaceUtil to API
