+++ modules/sdk/project-templates/project-templates/src/main/java/com/liferay/project/templates/ProjectTemplates.java
@@ -21,7 +21,6 @@ import com.liferay.project.templates.internal.Archetyper;
 import com.liferay.project.templates.internal.util.FileUtil;
 import com.liferay.project.templates.internal.util.StringUtil;
 import com.liferay.project.templates.internal.util.Validator;
-import com.liferay.project.templates.internal.util.WorkspaceUtil;
 
 import java.io.File;
 import java.io.InputStream;
