+++ modules/sdk/project-templates/project-templates/src/main/java/com/liferay/project/templates/WorkspaceUtil.java
@@ -12,7 +12,10 @@
  * details.
  */
 
-package com.liferay.project.templates.internal.util;
+package com.liferay.project.templates;
+
+import com.liferay.project.templates.internal.util.FileUtil;
+import com.liferay.project.templates.internal.util.StringUtil;
 
 import java.io.File;
 import java.io.IOException;
